refactor(web): give the actions popover's helpers a home
The popover carried its own constants, pure helpers, view-state type and
two hand-rolled fetches. They now sit where the rest of lib/tools lives,
and the file is 894 lines instead of 1059.

  constants.ts  tool descriptions, admin config links, tooltip suffixes
  utils.ts      buildTooltipMessage, getToolTooltip, getAdminConfigureInfo
  types.ts      SecondaryViewState

Two of those constants were declared inside `buildTooltipMessage`, so they
were rebuilt on every call.

`getConfiguredSources` existed twice — once in lib/hooks.ts and once in
the popover — written independently and disagreeing: the popover's version
drops `not_applicable` from the source list, the other kept it. The
popover's is the one that survives, in lib/sources.ts beside the
getSourceMetadata it calls.

The two credential submitters both rebuilt the request that
`saveMCPUserCredentials` already makes. They call it now. The old ones
hardcoded `"streamable-http"` where the service sends `"STREAMABLE_HTTP"`;
the endpoint upper-cases and swaps the dash, so both arrive as the same
enum member.
