+++ web/src/lib/tools/utils.ts
@@ -1,5 +1,13 @@
 import { SOURCE_METADATA_MAP } from "@/lib/sources";
-import { MethodSpec } from "@/lib/tools/types";
+import { MethodSpec, ToolSnapshot } from "@/lib/tools/types";
+import {
+  ADMIN_CONFIG_LINKS,
+  CONFIGURE_MESSAGE,
+  DEFAULT_TOOL_DESCRIPTION,
+  OPENAPI_ADMIN_CONFIG,
+  TOOL_DESCRIPTIONS,
+  USER_NOT_ADMIN_MESSAGE,
+} from "@/lib/tools/constants";
 import type { IconProps } from "@opal/types";
 import { SvgFileText, SvgServer } from "@opal/icons";
 
@@ -95,4 +103,45 @@ export function extractMethodSpecsFromDefinition(
   }
 
   return methods;
+}
+
+/**
+ * The tooltip for an action row: what the tool does, plus how to turn it on
+ * when it is not configured yet.
+ */
+export function buildTooltipMessage(
+  actionDescription: string,
+  isConfigured: boolean,
+  canManageAction: boolean
+): string {
+  if (isConfigured) return actionDescription;
+  return canManageAction
+    ? `${actionDescription} ${CONFIGURE_MESSAGE}`
+    : `${actionDescription} ${USER_NOT_ADMIN_MESSAGE}`;
+}
+
+/** {@link buildTooltipMessage}, with the description resolved from the tool. */
+export function getToolTooltip(
+  tool: ToolSnapshot,
+  isConfigured: boolean,
+  canManageAction: boolean
+): string {
+  const description =
+    (tool.in_code_tool_id && TOOL_DESCRIPTIONS[tool.in_code_tool_id]) ||
+    tool.description ||
+    DEFAULT_TOOL_DESCRIPTION;
+  return buildTooltipMessage(description, isConfigured, canManageAction);
+}
+
+/** Where an admin configures this tool, or null when there is nowhere to go. */
+export function getAdminConfigureInfo(
+  tool: ToolSnapshot
+): { href: string; tooltip: string } | null {
+  if (tool.in_code_tool_id && ADMIN_CONFIG_LINKS[tool.in_code_tool_id]) {
+    return ADMIN_CONFIG_LINKS[tool.in_code_tool_id] ?? null;
+  }
+  if (!tool.in_code_tool_id && !tool.mcp_server_id) {
+    return OPENAPI_ADMIN_CONFIG;
+  }
+  return null;
 }
